Set up components for working at the REPL
 - Add user.clj, which allows me to start and stop the system during
   development.

 - Add main.clj, where the whole system is constructed.

 - Add web_server.clj, which is routes and the Jetty for testing.

This is what I made out of the things I found on the web, which are
named in the Credits section in each file. That was complicated enough,
so I didn't think about deployment. Fitting the whole thing into a WAR
file will require some restructurings, I guess. Especially routes and
server in one file…