Extract routes into a separate component
I have to serve two targets now: development, where I start Jetty at the
REPL and deployment, where a Tomcat (probably) gets a WAR file. Both
need the routing information/a handler function, so it can't stay with
one of them. Therefore take it out of the file with the Jetty stuff and
put it in a separate one.

I also put it into its own component, because it appears to make sense.
But I don't understand the component stuff completely, yet. So I'm not
sure about its proper use.