Prepare for use with Tomcat
On http://thinkrelevance.com/blog/2013/06/04/clojure-workflow-reloaded
and in the video about the component framework (both referenced in the
files) there's some vague information about how to set up the system, so
that it can be deployed to something like Tomcat as a WAR file. And also
how to make the handler non-static.

Puzzle together these information in order to deploy this application to
my local Tomcat. I'm not sure whatsoever if I've put the puzzle together
in the right way. But that will show soon enough when I have to add more
components. Also add the necessary build configuration and some
documentation.