Quickly stitch together page generation
Right now I'm focusing on designing the page. I'm writing templates in
HTML and will start writing CSS with Garden. Add some Clojure wiring, so
that I can view the results in the browser. This is temporary, however,
and will be restructured when I'm reasonably content with the look and
feel.