welcome_page: Move out and add font declaration
This is not as it should be: refactoring and adding stuff in one. But
the addition was so small… Anyway, put the font declaration stuff from
welcome_page.clj in its own file. Add a font declaration for DejaVu
Sans. I'm probably not going to use this one, but it probably has the
same format as all other declarations of fonts from Font Squirrel.