Make JavaScript really work
JavaScript didn't work in the browser and I had a hard time figuring out
what exactly I had done differently compared to David Nolen's code. I am
not completely sure, but I think the reason was that the REPL in
scripts/repl.clj had the same :output-dir as the lein cljsbuild. Because
the JavaScript in the browser always worked until I fired up
scripts/repl. Therefore give them separate :output-dir's. Note that
scripts/repl.clj and scripts/brepl.clj still have the same :output-dir.
This should be correct, because the lein mies template also has it.

The other problem was that the paths to the JavaScript files in
try_it_out.html were relative paths, so the browser tried to find
JavaScript in /html/js/…, which of course went wrong. Fix that.