Bash everything into place for om-sync
I'm probably going to use om-sync for a large part of the client-server
communication. So set up some test code for that. The test code actually
got bigger than I expected. – om-sync is supposed to be a component that
can be installed independently from the other Om stuff. – And it's not
working completely alright. – PUT request seem both to fail and to
succeed. – But my weekly time is running out, so I'll quickly save my
state here.