Make the input field get focus
It was rather annoying always having to click into that field. So add
the capability of autofocusing. (It appears that React doesn't support
the corresponding HTML5 attribute.)

The other change is just a minor rearrangement.

I think I'm understanding well enough now how it works. Time for setting
up the database stuff!