Add Datomic database component for development
Add the dependencies and all it takes to connect to a local Datomic
database for local development.

I didn't update the deployment stuff yet, since I'll have to think about
that anyway. And there's a bug with the lein ring uberwar build, which
I'll have to take care of. (See the CI results.) I'm going to fix these
things next week.