Persist input from browser. Refactor components.
I had two things in place: The user input from the text field in the
browser was sent to the backend. The backend could set up a database
connection and load the schema if necessary. Put these things together
by storing the user input in the database and sending up the last input
when the browser reloads.

I did a second thing in this commit: Refactor how the Routes component
is created and used by the WebServer component. The original version was
based on a misunderstanding of how components should be used. But Stuart
Sierra writes that you should »Define the functions implementing the
behavior of the component to take an instance of the component as an
argument«. So that's what we do now.

Sorry for committing them together. I had written them both at the same
time and the text wasn't independent, so that I couldn't just do `git
add -p`.