Rebuild input/submit test with Kioo
This is structurally the same as before with Om only. But change the
document generation with Om functions to the scrape-and-substitute
approach from Kioo/Enlive.

The template document I added contains some stuff from Bootstrap. This
is because I plan to use Kioo and Bootstrap (not Om Bootstrap) together.
I thought that I would get Bootstrap styling right away, but Kioo only
uses the body part of the document, so this didn't quite work. It should
be easy to add though.

There are still some issues. I couldn't use a label with "for"
attribute, because React wants some htmlFor thing and Kioo doesn't take
that into account. (At least that's what I suspect.) Also I wanted to
put the input field and the button into a form, but I couldn't prevent
the browser from doing the submit form thing.

And in the code everything is mashed into one place. I have to find out
how exactly I'm supposed to use Kioo and Om in order to get a nice
separation.