Pull apart components
Where previously everything was built up in the io-view component, make
this component call out to form-view for creating the input field and
button part of the test page. I tried to define these things with
kioo/def(template|snippet), but that didn't work. For some reason the
passed-in owner wasn't right. Whatever. Back to kioo/component.