Add a real schema
I don't know a lot about databases, so there might be some issues with
this schema. It's just the easiest thing I could come up with for a
start. I'm especially unsure whether to put tags as plain strings or
separate entities and whether to make them user-owned or system-wide.
The way I'm doing it now is mostly gut-feeling and who knows what that
is worth.