Add primitive search and response
Remove all the last-input stuff from routes.clj and add a query route
that performs a full-text search on the database and returns the result.
I intend to support full-text search on titles, tags and URIs at once,
but keep it small for now, supporting only titles.

Add a template that displays a search field and a result list. I wanted
the search field and the button nicely joined together and copied the
necessary snippet from the Bootstrap documentation. Unfortunatly it
doesn't work. I suspect this is because some part of my React setup
(Kioo?) scatters empty spans in between elements. But this is not
important, therefore I won't investigate it now.

Remove the om-sync stuff from core.cljs and set up the user interface
with search field and result list and requesting search results. A few
things to say here:

 - I'm not sure if om-sync will make sense in my case. (Or maybe I just
   don't know how to use it properly.) For now things appear to be going
   in a different direction, so throw out all the om-sync stuff
   and leave only edn-xhr for convenience. There probably is a more
   appropriate wrapper for XHR, though.

 - Put in some channels. These are from the Om Advanced Tutorial and I'm
   not sure if I'll need them, but I leave them there, just in case. I
   considered using the Request Channel pattern, but then came to the
   conclusion that it just adds a level of indirection, which I don't
   need for now. Therefore keep things small and call process-input
   directly. (Instead of putting a request on the channel and having a
   request handler call the equivalent of process-input.)

 - The ref-cursor pattern is also from the Om Advanced Tutorial. I'm not
   sure whether using it is appropriate in my case or not. There are too
   many choices and too few guidelines for passing around data in Om.

 - I'm often passing nil to om/build. This is a result of the ref-cursor
   pattern: you don't pass around cursors anymore. I'll have to see if
   there is a shortcut to om/build without a second argument or
   something like that. Maybe I'm just misunderstanding some part of the
   system.

 - Search strings are not treated in any way, but there seems to be some
   sort of encoding and decoding going on under the hood, so that spaces
   and other not URI-conform characters come out well at the other end.
   However, that's not what other search engines do. I'll have to think
   about how to do this properly.

Sorry for the large commit, but if I split it into smaller pieces, the
system wouldn't have worked in between and that's also not optimal.