Update ClojureScript version, add notifications
Just noticed that we have a new ClojureScript release. (I probably
missed some in between.)

Also add a hook that sends me a notification whenever the cljs autobuild
finishes. It's not ideal, though, since it says Success even if there
are warnings. Also the command is not platform-independent. Just tell me
if you are on a non-Linux system and want to collaborate.