core.cljs: Do some cleanup
Rearrange forms. Add headings. Just for better overview.

Make focusing on the search input field automatic by putting it in the
template. For some reason that doesn't work yet. – Apparently the
autofocus attribute does make its way to the browser. I'll have to fix
that.

Staging for this commit was some wild git add -p thing. I hope that the
code still compiles and works after this commit. If not, it will after
the next one.