schema and sample data: Add headings
I have read that source code layout is important, so I'm starting to not
just dump my functions one after the other, but introduce a logical
order and sectioning. This already began in an earlier commit. Do it now
also for the transaction data.

Note that I won't correct this in all files at once. I'll do it
gradually and only where I see it fit. There is no point in paying
attention to decorations when the code hasn't solidified yet.