Fix Kioo dependency
The project depended on a locally deployed version of Kioo and Leiningen
emitted warnings about version ranges. Now Creighton uploaded the
current development version of Kioo to Clojars. Make the dependency in
project.clj fixed to that particular snapshot for reproducibility. Also
add an :exclusions clause according to Leiningen's recommendation.