Add first part of system overview
I decided to write the system overview as a series of stories that tell
what is going on under the hood in different scenarios. Add the
introductory paragraph for the system overview and the first story.

In an earlier commit I had changed the markup of link titles that are
file paths. It looks a bit nicer in the browser, but quite cluttered in
the editor. All in all it's not worth the effort, so don't continue
doing it.