Change exception handling, separate layers
Originally I only wanted to tidy up the test code for the transaction
data generators, but this led me to quite extensive thinking about the
layering of the system and separation of concerns, which in turn made me
reconsider exception handling. So I changed a bunch of things, but
didn't tidy up the test code, yet. :-)

Layering… I had already noticed before that routes/add-material (née
routes/save-material) contained a lot of stuff that doesn't belong in
routes. I decided that the storage system should be very open to
read-accesses, but very protective around write-accesses. That means,
everyone should be able to run raw datomic.api/q, but transactions
putting stuff into the database should be hidden behind an interface of
procedures (SICP terminology) for write-access. This way we ensure that
everything in the database is correct (at least structurally), but also
that reading code doesn't have to do awkward stuff in order to make
complex queries.

Put the interface of procedures for write-access into
database/transactions. This might only be temporary, since, looking
forward, the Datomic database probably won't be the only thing we need
to write to. There's also a search engine and a document store looming.
However, it should be easy to put a façade in front of the storage layer
later, so don't put a seemingly delegate-only namespace right now.

Making layering clearer also raises the question where to handle
exceptions. I searched the web and happened upon Dire, Slingshot and the
Conditions and Restarts chapter of Peter Seibel's Practical Common Lisp.
While the latter definitely helped me thinking about dealing with
exceptions (or conditions in general), Dire and Slingshot are all that
is quickly available for Clojure right now. And it's not too little. – I
think they can be used to construct something like the Common Lisp thing
when that is required. For now I like the separation of exception
handling and regular code. I don't like things with extra behaviours
that are not visible, though. Therefore use the without-bang,
with-supervise flavour of Dire. That might become a little clumsy,
though, so I'll have to see.

Oh, but I was writing about _where_ to handle exceptions. This might be
a bit surprising, but add exception handling at two levels. Of course we
could handle the ExecutionException in only one place, the last point
where it can be handled, namely in routes/add-material (or its handler,
respectively). However, letting an exception from the database layer
percolate up to the routing layer doesn't seem sensible, since the
routing layer shouldn't have to know about the internals of the database
layer. But an ExecutionException is definitely a very internal thing.
Still, the database layer has to signal that something went wrong.
Therefore catch the original exception and throw something else instead.

We could just return a special return value, but that wouldn't be right
either. It would be complecting return values with exceptions. There is
a reason people put exception handling mechanisms into the language, so
we should use them.