Add logging
There is a problem with adding materials, but the println I put in the
handler isn't showing at the REPL. Add logging so that it goes to a
file. (Of course I'm expecting to make more use of logging in the
future. Otherwise I would just turn my println into a spit.)

Put logging configuration in an extra component and add it as a
dependency to the database component. The weird thing is, that the
database never accesses the injected object. Instead, logging is a
global state thing and thus doesn't need to be passed around. I'm not
sure whether this is good, but what I've read on the web all says that
logging is one of the few exceptions where global state is okay. Well,
at least it would be really annoying to pass a logger down to all
functions that need to log something.