Fix arguments in database/add-material
Actually, the problem I talked about in the parent commit didn't have to
do with the ExecutionException I'm handling in database.transactions. It
was wrong function arguments. Thus, logging as I installed it wasn't
helpful, but actually timbre/spy was quite useful. The strange thing is
that the IllegalArgumentException or what it was never became visible.
It was just ignored. :-(

Also remove a comment about namespacing condition types. Namespacing
might come later, though.