Try to emulate re-frame in Om
I read through the re-frame documentation and found it appealing. So
appealing that I want Theatralia to follow that pattern. I tried to
coerce Om into using it and it nearly succumbed. However, the concept of
signals cannot be expressed in a sensible way. – We have to thread the
owner through the call chain and this can get rather ugly. – Some more
of my considerations are written down in the big comment. So, I'll have
to abandon my Om and look for new one.