Thomsky: Port subscriptions from re-frame
I decided to put the app in a record instead of making everything
global. I don't know why re-frame makes everything global. I mean, it's
more convenient in a way, but also brings with it all the problems of
globality. For example, before I decided to port the whole of re-frame,
I considered replacing the global app-db atom with a Datascript atom and
just adding different middleware. Don't know if that would have worked,
but switching out the kind-of-private global app-db would have been
ugly.

Introducing interfaces is a side-effect of using a record. Maybe it
brings some other benefits. I wanted to define the implementations in
the same file as the interfaces are defined (definitely not all
implementations in a giant record), but I guess that would be a bad idea
as well. I'm not quite content with the code organization right now, but
I'll see.