Use Datascript with re-frame
As indicated in the previous commit, all the work porting re-frame to
Datascript was unnecessary. re-frame works quite well with Datascript.

For nostalgic reasons use Thomsky as name of the namespace for the
Datascript adaptor stuff. Add two macros that make the registration
easier. Man, ClojureScript documentation and tooling sucks! Or is it me?
Another sucking point are the reflective capabilities: I wanted to
implement the registration sugar as procedures, but there is no var-get…
And the other possibilities I found were also quite horrible.

theatralia.core now only contains some sample code that I used to try
out the Datascript integration.