Thomsky: allow parameterized queries
Dave Dixon's `bind` only put @conn as the query's parameter. This is
rather limiting. Add the possibility to provide arbitrary parameters. In
order to make the code (and my mind) clearer, remove the `state`
argument from `bind`. I didn't see its necessity for now and it clashed
with the query parameters.