Client: Correct problems with scratch area
For some hysteric reason storing a map as an attribute value in
Datascript stopped working.
https://github.com/tonsky/datascript/issues/69 is the reason, I think,
but shouldn't it be fixed according to that discussion? And why did it
work before and doesn't now? It makes no sense.

(Yes, I did bump to Datascript 0.11.5, but when I debumped and cleaned
several times, it didn't start working again.)

Anyway, restructure some things, so that I don't need to store maps
anymore. Some refactorings strewn in as usual. Sorry for that.