Client: Work around limitation of bind
I discovered strange behaviour in the client:

 1. Search for "food". – It shows the expected result.
 2. Search for "foo". – It shows the expected empty result.
 3. Add "*", so that it becomes "foo*". – The result stays empty,
    although the first result should be shown again.

(Behaviour for when only the sample data are loaded.)

The problem was that I used a "single scalar" ([:find ?rs . …]) find
specification in my query for search results: thomsky/bind looks at the
:tx-data. When going from an empty search result to a non-empty search
result, :tx-data contains two Datoms. In the first the empty result is
de-asserted, in the second the new non-empty result is asserted.

 [#datascript/Datom [… nil false]
  #datascript/Datom [… [["On Food and Cooking"] ["English Food"]] true]]

thomsky/bind runs the query over these two Datoms, and, because it's
"single scalar", it only returns the nil value for the first Datom. The
value for the second Datom is ignored. So the "novelty" tests empty and
the result Ratom is not updated.

Work around this by making the whole thing a bit databasier: don't store
a vector as a single attribute value of the search result entity, but
make the attribute multi-valued and store each element of the result in
a separate attribute value. Of course, in the future we have to go even
further, because we'll want to query other attributes of the individual
search result, for example the score. But for now the vectors are good
enough. Note that the by-score sort the server did is lost, because we
have to feed the results to the transactor as a set.

There is still a query with a "single scalar" find specification out
there: the query for the scratch entity ID. This is fine, because we're
not retracting the scratch entities completely. However, the whole can't
stay like that: either we warn about the case of "single scalars" or we
put in place some primitive query parsing that detects them and changes
the behaviour of the no-novelty detection.