Add crude test if a query is single-valued or not
If we can determine if a query returns a single value and not a
collection of results, we can use thomsky/bind on single-valued queries
again. This is the first step: add a single-valued? predicate and a
test. Next I have to set up the test environment and write more test
cases.