CLJS tests: Hide reagent.core/render call
As reported in bensu/doo#16, lein doo slimer didn't work, and as
correctly diagnosed by @bensu, the reason was that reagent.core/render
was called during testing, which doesn't make sense.

Wrap the code that starts the application in a procedure and call this
procedure from code that doesn't get loaded in tests.