Sandbox/tag input: Throw out empty tag
We want a list of non-empty tag inputs and an empty tag input that
spawns an empty tag input when someone edits it. The old approach of
having separate divs and separate Kioo transformations didn't work,
because I couldn't bind it to the database cleanly and keep focus on the
right input fields when the new tag input was added.

I solved it by adding an empty tag to the list of tags retrieved from
the database and changing the callbacks a bit. There is still a problem,
which will be fixed in one of the upcoming commits.

Throw out some stuff that became obsolete with the new solution.