tag inputs: Improve behaviour
Introduce a new logic for determining when :new-tag/:remove-tag events
should be dispatched. – Not from onFocus/onBlur callbacks, but from
changes of the tag input field. Now we only put a tag in the database
when it becomes non-empty and remove it as soon it gets empty.

This logic is more simple and the resulting behaviour is less annoying.
Tabbing works out of the box now. However, there are still slight
quirks. Leave TODOs for them.