C subs views: Tidy up subscription IDs
This is an attempt at making the client more orderly. Bring some
regularity into the IDs of subscriptions. I'm pretty clueless about how
the overall thing is going to be structured. Right now it appears to be
a small ball of mud. Is there a good book about UI programming? Or does
the centre of the problem lie somewhere else?